Create separate files for all definitions
Issues: #82
