+++ SteamScout/Models/Definitions/PropertyListReadable.swift
@@ -1,0 +1,14 @@
+//
+//  PropertyListReadable.swift
+//  SteamScout
+//
+//  Created by Srinivas Dhanwada on 3/19/17.
+//  Copyright © 2017 dhanwada. All rights reserved.
+//
+
+import Foundation
+
+protocol PropertyListReadable {
+    func propertyListRepresentation() -> NSDictionary
+    init?(propertyListRepresentation:NSDictionary?)
+}
