+++ SteamScout/Models/Definitions/UpdateType.swift
@@ -1,0 +1,13 @@
+//
+//  UpdateType.swift
+//  SteamScout
+//
+//  Created by Srinivas Dhanwada on 3/19/17.
+//  Copyright © 2017 dhanwada. All rights reserved.
+//
+
+import Foundation
+
+enum UpdateType : Int {
+    case none = 0, teamInfo, fieldSetup, finalStats, actionsEdited, autonomous, teleop
+}
